requirements.txt
Django==1.6.2
Pillow==2.3.0
South==0.8.4
argparse==1.2.1
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-classy-tags==0.4
-e git+https://github.com/divio/django-cms.git@3bc4a2535d634358854747fe95e04cf4e11e5ac2#egg=django_cms-dev
django-mptt==0.6.0
django-sekizai==0.7
django-toolbelt==0.0.1
djangocms-admin-style==0.2.0
djangocms-text-ckeditor==2.1.2
gunicorn==18.0
html5lib==1.0b3
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
pyuploadcare==1.2.3
requests==2.2.1
six==1.5.2
static==1.0.2
wsgiref==0.1.2
